Adjust properly the `textwidth` option.
We want a `textwidth=78` to show a column as a guide for long text. The
problem was that with a global setting, some plugins with customized
`textwidth` options (like `gitcommit`) ignore their own value and use
ours.

An autocommand was made to put a local `textwidth=78` only if there were
no previous value.